Improve Twig extension
